split modifiers into new file
